refactor(protocol): split policy module
