+++ packages/protocol/src/policy/point-registry.ts
@@ -1,0 +1,237 @@
+import { PolicyPointContractModule } from "./point-contract.js";
+
+export namespace PolicyPointRegistryModule {
+  const contract = PolicyPointContractModule.contract;
+  const preBoundary = PolicyPointContractModule.preBoundary;
+  const postBoundary = PolicyPointContractModule.postBoundary;
+  const Timing = PolicyPointContractModule.Timing;
+  type Timing = PolicyPointContractModule.Timing;
+  type RegisteredPolicyPointId = PolicyPointContractModule.RegisteredPolicyPointId;
+  type PolicyPointContract = PolicyPointContractModule.PolicyPointContract;
+
+  export const PolicyPointRegistry = {
+    "session.inbound.pre": contract(
+      "session.inbound.pre",
+      "pre",
+      ["session"],
+      ["actorId", "sessionId", "inboundEvent"],
+      ["audit.annotate", "run.abort", "delegation.set_constraints"],
+      ...preBoundary,
+    ),
+    "run.lifecycle.pre": contract(
+      "run.lifecycle.pre",
+      "pre",
+      ["run"],
+      ["actorId", "sessionId", "runId"],
+      [
+        "audit.annotate",
+        "run.abort",
+        "delegation.set_constraints",
+        "prompt.append_context",
+        "prompt.inject_message",
+      ],
+      ...preBoundary,
+    ),
+    "run.turn.pre": contract(
+      "run.turn.pre",
+      "pre",
+      ["run"],
+      ["sessionId", "runId", "turnIndex"],
+      [
+        "audit.annotate",
+        "run.abort",
+        "run.retry_after",
+        "prompt.append_context",
+        "prompt.inject_message",
+      ],
+      ...preBoundary,
+    ),
+    "prompt.context.pre": contract(
+      "prompt.context.pre",
+      "pre",
+      ["prompt"],
+      ["sessionId", "runId", "turnIndex"],
+      ["prompt.append_context", "prompt.inject_message", "prompt.replace", "audit.annotate"],
+      // Prompt context policy is advisory: failure should not block the run,
+      // but it still marks a side-effect boundary because it can rewrite prompt input.
+      "fail-open",
+      true,
+    ),
+    "tool.catalog.pre": contract(
+      "tool.catalog.pre",
+      "pre",
+      ["tool"],
+      ["sessionId", "runId", "availableTools"],
+      ["tool.filter", "audit.annotate", "run.abort"],
+      ...preBoundary,
+    ),
+    "connection.llm.pre": contract(
+      "connection.llm.pre",
+      "pre",
+      ["connection"],
+      ["sessionId", "runId", "modelId"],
+      ["prompt.append_context", "prompt.inject_message", "run.abort", "audit.annotate"],
+      ...preBoundary,
+    ),
+    "connection.llm.post": contract(
+      "connection.llm.post",
+      "post",
+      ["connection"],
+      ["sessionId", "runId", "modelId", "responseTokens"],
+      ["audit.annotate", "run.abort", "prompt.inject_message", "run.replace_messages"],
+      ...postBoundary,
+    ),
+    "tool.native.pre": contract(
+      "tool.native.pre",
+      "pre",
+      ["tool"],
+      ["sessionId", "runId", "toolId", "toolInput"],
+      [
+        "tool.filter",
+        "tool.rewrite_input",
+        "tool.skip_invocation",
+        "tool.require_approval",
+        "run.abort",
+        "audit.annotate",
+      ],
+      ...preBoundary,
+    ),
+    "tool.mcp.pre": contract(
+      "tool.mcp.pre",
+      "pre",
+      ["tool"],
+      ["sessionId", "runId", "toolId", "mcpServerId", "toolInput"],
+      [
+        "tool.filter",
+        "tool.rewrite_input",
+        "tool.skip_invocation",
+        "tool.require_approval",
+        "run.abort",
+        "audit.annotate",
+      ],
+      ...preBoundary,
+    ),
+    "delegation.subagent.pre": contract(
+      "delegation.subagent.pre",
+      "pre",
+      ["worker"],
+      ["sessionId", "runId", "subagentId", "subagentProfile"],
+      ["delegation.set_constraints", "delegation.require_approval", "run.abort", "audit.annotate"],
+      ...preBoundary,
+    ),
+    "delegation.background.pre": contract(
+      "delegation.background.pre",
+      "pre",
+      ["worker"],
+      ["sessionId", "runId", "backgroundTaskId"],
+      ["delegation.set_constraints", "delegation.require_approval", "run.abort", "audit.annotate"],
+      ...preBoundary,
+    ),
+    "tool.native.post": contract(
+      "tool.native.post",
+      "post",
+      ["tool"],
+      ["sessionId", "runId", "toolId", "toolResult"],
+      ["audit.annotate", "run.abort", "tool.rewrite_output"],
+      ...postBoundary,
+    ),
+    "tool.mcp.post": contract(
+      "tool.mcp.post",
+      "post",
+      ["tool"],
+      ["sessionId", "runId", "toolId", "mcpServerId", "toolResult"],
+      ["audit.annotate", "run.abort", "tool.rewrite_output"],
+      ...postBoundary,
+    ),
+    "delegation.subagent.post": contract(
+      "delegation.subagent.post",
+      "post",
+      ["worker"],
+      ["sessionId", "runId", "subagentId", "subagentResult"],
+      ["audit.annotate"],
+      ...postBoundary,
+    ),
+    "delegation.background.post": contract(
+      "delegation.background.post",
+      "post",
+      ["worker"],
+      ["sessionId", "runId", "backgroundTaskId", "taskResult"],
+      ["audit.annotate"],
+      ...postBoundary,
+    ),
+    "run.turn.post": contract(
+      "run.turn.post",
+      "post",
+      ["run"],
+      ["sessionId", "runId", "turnIndex", "turnResult"],
+      [
+        "audit.annotate",
+        "run.abort",
+        "run.continue_with_prompt",
+        "prompt.inject_message",
+        "run.replace_messages",
+      ],
+      ...postBoundary,
+    ),
+    "run.completion.pre": contract(
+      "run.completion.pre",
+      "pre",
+      ["run"],
+      ["sessionId", "runId", "completionCandidate"],
+      ["audit.annotate", "run.abort", "prompt.append_context", "run.replace_messages"],
+      ...preBoundary,
+    ),
+    "session.writeback.pre": contract(
+      "session.writeback.pre",
+      "pre",
+      ["session"],
+      ["sessionId", "runId", "writebackPayload"],
+      ["audit.annotate", "run.abort", "writeback.rewrite", "writeback.suppress"],
+      ...preBoundary,
+    ),
+    "run.lifecycle.post": contract(
+      "run.lifecycle.post",
+      "post",
+      ["run"],
+      ["sessionId", "runId", "runOutcome"],
+      ["audit.annotate"],
+      ...postBoundary,
+    ),
+    "run.error.error": contract(
+      "run.error.error",
+      "error",
+      ["run"],
+      ["sessionId", "runId", "errorCode", "errorPhase"],
+      ["audit.annotate", "run.abort", "run.retry_after"],
+      "fail-closed",
+      false,
+    ),
+  } satisfies Record<RegisteredPolicyPointId, PolicyPointContract>;
+
+  export const policyPointMigrationMapping = {
+    [Timing.INBOUND_RECEIVE]: ["session.inbound.pre"],
+    [Timing.RUN_START]: ["run.lifecycle.pre"],
+    [Timing.TURN_START]: ["run.turn.pre"],
+    [Timing.CONTEXT_PREPARE]: ["prompt.context.pre"],
+    [Timing.RESOURCES_PREPARE]: ["tool.catalog.pre"],
+    [Timing.MODEL_REQUEST]: ["connection.llm.pre"],
+    [Timing.MODEL_RESPONSE]: ["connection.llm.post"],
+    [Timing.INVOKE_PREPARE]: [
+      "tool.native.pre",
+      "tool.mcp.pre",
+      "delegation.subagent.pre",
+      "delegation.background.pre",
+    ],
+    [Timing.INVOKE_RESULT]: [
+      "tool.native.post",
+      "tool.mcp.post",
+      "delegation.subagent.post",
+      "delegation.background.post",
+    ],
+    [Timing.TURN_FINISH]: ["run.turn.post"],
+    [Timing.COMPLETION_PREPARE]: ["run.completion.pre"],
+    [Timing.WRITEBACK_COMMIT]: ["session.writeback.pre"],
+    [Timing.RUN_FINISH]: ["run.lifecycle.post"],
+    [Timing.ERROR]: ["run.error.error"],
+  } satisfies Record<Timing, RegisteredPolicyPointId[]>;
+}
