+++ packages/protocol/src/policy/version.ts
@@ -1,0 +1,1 @@
+export const policyKernelVersion = 1;
